Move link-checker-api stubbing into separate module
We will need to stub link-checker-api responses in future tests,
which will be easier to do if this is neatly encapsulated in its
own file.
