+++ spec/support/step_nav_steps.rb
@@ -1,4 +1,6 @@
 module StepNavSteps
+  include LinkChecker
+
   def setup_publishing_api
     stub_any_publishing_api_put_content
     stub_any_publishing_api_discard_draft
@@ -91,29 +93,22 @@ module StepNavSteps
   end
 
   def given_there_is_a_step_by_step_page_with_a_link_report
-    link_checker_api_get_batch(
-      id: 1,
-      links: [link_checker_api_link_report_success]
-    )
-
     step = create(:step)
-    create(:link_report, step_id: step.id)
+    stub_link_checker_report_success(step)
     @step_by_step_page = create(:step_by_step_page, steps: [step], slug: "step-by-step-with-link-report")
   end
 
   alias_method :given_there_is_a_step_that_has_no_broken_links, :given_there_is_a_step_by_step_page_with_a_link_report
 
   def given_there_is_a_step_by_step_page_with_unpublished_changes_whose_links_have_been_checked
-    link_checker_api_get_batch(id: 1, links: [link_checker_api_link_report_success])
     step = create(:step)
-    create(:link_report, step_id: step.id)
+    stub_link_checker_report_success(step)
     @step_by_step_page = create(:step_by_step_with_unpublished_changes, steps: [step], slug: 'step-by-step-with-unpublished-changes')
   end
 
   def given_a_step_by_step_has_been_updated_after_links_last_checked
-    link_checker_api_get_batch(id: 1, links: [link_checker_api_link_report_success])
     step = create(:step)
-    create(:link_report, step_id: step.id)
+    stub_link_checker_report_success(step)
     @step_by_step_page = create(
       :step_by_step_with_unpublished_changes,
       steps: [step],
@@ -123,7 +118,6 @@ module StepNavSteps
   end
 
   def given_a_step_by_step_has_an_empty_step_added_after_links_last_checked
-    link_checker_api_get_batch(id: 1, links: [link_checker_api_link_report_success])
     @step_by_step_page = create(
       :step_by_step_with_unpublished_changes,
       slug: 'step-by-step-with-link-report-and-empty-step-added-since-links-checked',
@@ -131,7 +125,7 @@ module StepNavSteps
     )
 
     step_with_link_report = create(:step, step_by_step_page: @step_by_step_page)
-    create(:link_report, step_id: step_with_link_report.id)
+    stub_link_checker_report_success(step_with_link_report)
     create(:step, contents: "", step_by_step_page: @step_by_step_page)
   end
 
@@ -160,40 +154,13 @@ module StepNavSteps
   def given_there_is_a_step_with_a_broken_link
     @step_by_step_page = create(:step_by_step_page)
     step = create(:step, step_by_step_page: @step_by_step_page)
-    link_checker_api_get_batch(
-      id: 1,
-      links: [link_checker_api_link_report_fail]
-    )
-    create(:link_report, step_id: step.id)
+    stub_link_checker_report_broken_link(step)
   end
 
   def given_there_is_a_step_with_multiple_broken_links
     @step_by_step_page = create(:step_by_step_page)
     step = create(:step, step_by_step_page: @step_by_step_page)
-    link_checker_api_get_batch(
-      id: 1,
-      links: [link_checker_api_link_report_fail, link_checker_api_link_report_fail]
-    )
-    create(:link_report, step_id: step.id)
-  end
-
-  def link_checker_api_link_report_success
-    {
-      "uri": "https://www.gov.uk/",
-      "status": "ok",
-      "checked": "2017-04-12T18:47:16Z",
-      "errors": [],
-      "warnings": [],
-      "problem_summary": "null",
-      "suggested_fix": "null"
-    }
-  end
-
-  def link_checker_api_link_report_fail
-    {
-      "uri": "https://www.gov.uk/foo",
-      "status": "broken"
-    }
+    stub_link_checker_report_multiple_broken_links(step)
   end
 
   def then_I_can_see_a_success_message(message)
